# docs/generated/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "1BIS Backend API",
    "version": "1.0.0",
    "description": "Welcome to the 1BIS API documentation. This is a RESTful API that allows you to interact with the 1BIS backend services.",
    "contact": {
      "name": "1BIS Support",
      "email": "support@1bis.com",
      "url": "https://1bis.com/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://1bis.com/terms"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Development server"
    },
    {
      "url": "https://api.1bis.com/api",
      "description": "Production server"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key for authentication"
      }
    },
...
  },
  "paths": {
...
    "/api/cartItem/{id}": {
      "get": {
        "summary": "Get cartItem by ID",
        "description": "Retrieve a specific cartItem by its unique identifier with optional field selection",
        "tags": [
          "CartItem"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$"
            },
            "description": "CartItem ID (MongoDB ObjectId format)",
            "example": "507f1f77bcf86cd799439011"
          },
          {
            "in": "query",
            "name": "fields",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated list of fields to include (supports nested fields with dot notation)",
            "example": "id,name,description,type"
          }
        ],
        "responses": {
          "200": {
            "description": "CartItem retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "cartItem": {
                              "$ref": "#/components/schemas/CartItem"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "patch": {
        "summary": "Update cartItem",
        "description": "Update cartItem data by ID (partial update)",
        "tags": [
          "CartItem"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$"
            },
            "description": "CartItem ID (MongoDB ObjectId format)",
            "example": "507f1f77bcf86cd799439011"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "minProperties": 1,
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "description": "CartItem name",
                    "example": "Updated Email CartItem"
                  },
                  "description": {
                    "type": "string",
                    "description": "CartItem description",
                    "example": "Updated description for the cartItem"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "email",
                      "sms",
                      "push",
                      "form"
                    ],
                    "description": "CartItem type for categorization",
                    "example": "email"
                  },
                  "isDeleted": {
                    "type": "boolean",
                    "description": "Soft delete flag",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "CartItem updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "cartItem": {
                              "$ref": "#/components/schemas/CartItem"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "Delete cartItem",
        "description": "Permanently delete a cartItem by ID",
        "tags": [
          "CartItem"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$"
            },
            "description": "CartItem ID (MongoDB ObjectId format)",
            "example": "507f1f77bcf86cd799439011"
          }
        ],
        "responses": {
          "200": {
            "description": "CartItem deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "description": "Empty object for successful deletion"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  }
}